# openapi.yaml
openapi: 3.0.3
info:
  title: LangGraph Multi-Agent System API
  description: |
    A comprehensive multi-agent workflow system built with FastAPI and LangGraph.

    This API provides endpoints for managing tasks, agents, workflows, and system operations.
    The system supports multiple execution modes, real-time communication via WebSockets,
    and comprehensive monitoring and health checking capabilities.

    ## Features
    - Task management with priority and type classification
    - Multi-agent coordination and execution
    - Real-time status updates via WebSocket
    - Comprehensive system monitoring and health checks
    - LLM configuration management
    - Batch operations and statistics
    - Meta-agent conversation management

    ## Authentication
    The API uses JWT-based authentication for securing endpoints.

    ## Rate Limiting
    Rate limiting is applied to prevent abuse and ensure system stability.

    ## WebSocket Support
    Real-time communication is supported through WebSocket endpoints for live updates.
  version: 1.0.0
  contact:
    name: LangGraph Multi-Agent Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/workflows/{workflow_id}/status:
    get:
      tags:
        - Workflows
      summary: Get workflow status
      description: Retrieve current status and progress of a specific workflow
      operationId: getWorkflowStatus
      parameters:
        - name: workflow_id
          in: path
          required: true
          description: Workflow ID
          schema:
            type: string
      responses:
        '200':
          description: Workflow status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStatus'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/workflows/{workflow_id}/config:
    put:
      tags:
        - Workflows
      summary: Update workflow configuration
      description: Update configuration parameters for a specific workflow
      operationId: updateWorkflowConfig
      parameters:
        - name: workflow_id
          in: path
          required: true
          description: Workflow ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowConfigRequest'
      responses:
        '200':
          description: Workflow configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/system/status:
    get:
      tags:
        - System
      summary: Get system status
      description: Retrieve basic system status information
      operationId: getSystemStatus
      responses:
        '200':
          description: System status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy, degraded, unhealthy, maintenance]
                  uptime_seconds:
                    type: integer
                  version:
                    type: string
                  environment:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/system/maintenance:
    post:
      tags:
        - System
      summary: Trigger maintenance
      description: Trigger system maintenance operations
      operationId: triggerMaintenance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - operation
              properties:
                operation:
                  type: string
                  enum: [cleanup, backup, optimize, reset_cache, health_check]
                parameters:
                  type: object
                  description: Operation-specific parameters
                force:
                  type: boolean
                  default: false
                  description: Force operation even if system is busy
      responses:
        '200':
          description: Maintenance operation triggered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation_id:
                    type: string
                  operation:
                    type: string
                  status:
                    type: string
                    enum: [queued, running, completed, failed]
                  started_at:
                    type: string
                    format: date-time
                  estimated_duration:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    WorkflowConfigRequest:
      type: object
      properties:
        execution_mode:
          $ref: '#/components/schemas/WorkflowExecutionMode'
        checkpoint_interval:
          type: integer
          minimum: 10
          maximum: 3600
          description: Checkpoint interval in seconds
        max_retries:
          type: integer
          minimum: 0
          maximum: 10
          description: Maximum retry attempts
        timeout_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          description: Timeout in seconds
        agents_config:
          type: object
          additionalProperties: true
          description: Agent configuration
...
    WorkflowExecutionMode:
      type: string
      enum: [sequential, parallel, adaptive]
      description: Workflow execution modes
...
    WorkflowStatus:
      type: object
      required:
        - workflow_id
        - task_id
        - current_phase
        - status
        - progress
        - active_agents
        - completed_agents
        - failed_agents
      properties:
        workflow_id:
          type: string
          description: Workflow ID
        task_id:
          type: string
          description: Task ID
        current_phase:
          type: string
          description: Current workflow phase
        status:
          type: string
          description: Workflow status
        progress:
          type: number
          minimum: 0
          maximum: 1
          description: Progress percentage
        active_agents:
          type: array
          items:
            type: string
          description: Active agents
        completed_agents:
          type: array
          items:
            type: string
          description: Completed agents
        failed_agents:
          type: array
          items:
            type: string
          description: Failed agents
        execution_time:
          type: number
          nullable: true
          description: Execution time in seconds
        estimated_remaining_time:
          type: number
          nullable: true
          description: Estimated remaining time in seconds
...
    ApiResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Response message
        data:
          type: object
          additionalProperties: true
          nullable: true
          description: Response data
        error_code:
          type: string
          nullable: true
          description: Error code if applicable
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      type: object
      required:
        - success
        - error_code
        - error_message
      properties:
        success:
          type: boolean
          default: false
          description: Always false for error responses
        error_code:
          type: string
          description: Machine-readable error code
        error_message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: BAD_REQUEST
            error_message: Invalid request parameters
            timestamp: "2024-01-01T00:00:00Z"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: UNAUTHORIZED
            error_message: Authentication required
            timestamp: "2024-01-01T00:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: NOT_FOUND
            error_message: Resource not found
            timestamp: "2024-01-01T00:00:00Z"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: INTERNAL_ERROR
            error_message: Internal server error
            timestamp: "2024-01-01T00:00:00Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
security:
  - BearerAuth: []
...