# src/langgraph_multi_agent/api/routes/system.py
"""系统管理API路由"""

import logging
from typing import Dict, Any
from datetime import datetime

from fastapi import APIRouter, HTTPException

from ..models import SystemHealth, SystemMetrics, ApiResponse

logger = logging.getLogger(__name__)

router = APIRouter()


@router.get("/health", response_model=SystemHealth)
async def get_system_health():
    """获取系统健康状态"""
    try:
        # 模拟系统健康数据
        system_health = SystemHealth(
            overall_healthy=True,
            timestamp=datetime.now(),
            health_checks={
                "database": {
                    "status": "healthy",
                    "response_time": 0.05,
                    "last_check": datetime.now().isoformat()
                },
                "redis": {
                    "status": "healthy",
                    "response_time": 0.02,
                    "last_check": datetime.now().isoformat()
                },
                "workflow_engine": {
                    "status": "healthy",
                    "active_workflows": 5,
                    "last_check": datetime.now().isoformat()
                }
            },
            active_alerts={},
            error_metrics={
                "total_errors": 0,
                "error_rate": 0.0,
                "last_error": None
            },
            performance_metrics={
                "cpu_usage": 45.2,
                "memory_usage": 62.8,
                "disk_usage": 35.1
            }
        )
        
        return system_health
        
    except Exception as e:
        logger.error(f"获取系统健康状态失败: {e}")
        raise HTTPException(status_code=500, detail=f"获取系统健康状态失败: {str(e)}")


@router.get("/metrics", response_model=SystemMetrics)
async def get_system_metrics():
    """获取系统指标"""
    try:
        system_metrics = SystemMetrics(
            cpu_usage=45.2,
            memory_usage=62.8,
            active_workflows=5,
            active_agents=8,
            requests_per_minute=120.5,
            average_response_time=0.25,
            error_rate=0.02,
            uptime_seconds=86400
        )
        
        return system_metrics
        
    except Exception as e:
        logger.error(f"获取系统指标失败: {e}")
        raise HTTPException(status_code=500, detail=f"获取系统指标失败: {str(e)}")


@router.get("/status", response_model=ApiResponse)
async def get_system_status():
    """获取系统状态"""
    try:
        return ApiResponse(
            success=True,
            message="系统状态正常",
            data={
                "status": "running",
                "database": {"connected": True, "response_time": 0.05},
                "redis": {"connected": True, "response_time": 0.02},
                "workflow_engine": {"active": True, "workflows": 5},
                "timestamp": datetime.now().isoformat()
            }
        )
        
    except Exception as e:
        logger.error(f"获取系统状态失败: {e}")
        raise HTTPException(status_code=500, detail=f"获取系统状态失败: {str(e)}")


@router.post("/maintenance", response_model=ApiResponse)
async def trigger_maintenance():
    """触发系统维护"""
    try:
        return ApiResponse(
            success=True,
            message="系统维护已触发",
            data={
                "maintenance_started": datetime.now().isoformat(),
                "estimated_duration": "10 minutes"
            }
        )
        
    except Exception as e:
        logger.error(f"触发系统维护失败: {e}")
        raise HTTPException(status_code=500, detail=f"触发系统维护失败: {str(e)}")